Move class TaskList into package duke.task
