+++ src/main/java/duke/Ui.java
@@ -1,5 +1,7 @@
 package duke;
 
+import duke.task.TaskList;
+
 import java.util.Scanner;
 
 public class Ui {
