+++ src/main/java/duke/command/ByeCommand.java
@@ -1,9 +1,9 @@
 package duke.command;
 
 import duke.ExitException;
-import duke.TaskList;
 import duke.Ui;
 import duke.Storage;
+import duke.task.TaskList;
 
 public class ByeCommand extends Command {
