+++ src/main/java/duke/command/FindCommand.java
@@ -1,9 +1,9 @@
 package duke.command;
 
 import duke.Storage;
-import duke.TaskList;
 import duke.Ui;
 import duke.task.Task;
+import duke.task.TaskList;
 
 import java.util.ArrayList;
