+++ src/main/java/duke/command/UnrecognizedCommand.java
@@ -1,8 +1,8 @@
 package duke.command;
 
-import duke.TaskList;
 import duke.Ui;
 import duke.Storage;
+import duke.task.TaskList;
 
 public class UnrecognizedCommand extends Command {
